+++ src/character/personality.rs
@@ -3,22 +3,21 @@ use serde::{Deserialize, Serialize};
 
 use super::quality::Quality;
 
-const RNG_VARIANCE: i32 = 5;
+const RNG_VARIANCE: i16 = 5;
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct Personality {
-    pub optimism: i32,
-    pub altruism: i32,
-    pub reflexes: i32,
-    pub teamwork: i32,
-    pub awareness: i32,
-    pub focus: i32,
+    pub optimism: i16,
+    pub altruism: i16,
+    pub reflexes: i16,
+    pub teamwork: i16,
+    pub awareness: i16,
+    pub focus: i16,
 }
 
-fn get_variance<R: Rng + ?Sized>(rng: &mut R) -> i32 {
+fn get_variance<R: Rng + ?Sized>(rng: &mut R) -> i16 {
     rng.gen_range(-RNG_VARIANCE..=RNG_VARIANCE)
 }
-
 
 pub fn get_personality<R: Rng + ?Sized>(rng: &mut R, quality: Quality) -> Personality {
     Personality {
